Fix reported installed size change with big packages (Closes: #759769)
* Fix reported installed size change when installing/deinstalling packages
  which are bigger than 2 GiB (int type size 32 bits).  Thanks Benny Baumann
  and Jason Rhinelander for the report and info, and Sven Joachim for the
  patch.  (Closes: #759769)

+++ src/cmdline/cmdline_prompt.cc
@@ -352,7 +352,7 @@ static void cmdline_show_instinfo(pkgvector &items,
       // Show the change in size between the versions.
       if(showsize)
 	{
-	  int dsize=(instver.end()?0:instver->InstalledSize)
+	  long long dsize=(instver.end()?0:instver->InstalledSize)
 	    -(i->CurrentVer().end()?0:i->CurrentVer()->InstalledSize);
 
 	  if(dsize>0)
